refactor: move link to vo package
