+++ src/main/java/com/meoguri/linkocean/domain/linkmetadata/persistence/FindLinkMetadataByUrlQuery.java
@@ -3,8 +3,8 @@ package com.meoguri.linkocean.domain.linkmetadata.persistence;
 import static java.lang.String.*;
 
 import com.meoguri.linkocean.annotation.Query;
-import com.meoguri.linkocean.domain.linkmetadata.entity.Link;
 import com.meoguri.linkocean.domain.linkmetadata.entity.LinkMetadata;
+import com.meoguri.linkocean.domain.linkmetadata.entity.vo.Link;
 import com.meoguri.linkocean.exception.LinkoceanRuntimeException;
 
 import lombok.RequiredArgsConstructor;
