+++ src/test/java/com/meoguri/linkocean/domain/linkmetadata/entity/vo/LinkTest.java
@@ -1,4 +1,4 @@
-package com.meoguri.linkocean.domain.linkmetadata.entity;
+package com.meoguri.linkocean.domain.linkmetadata.entity.vo;
 
 import static org.assertj.core.api.Assertions.*;
